@@ -33,15 +43,32 @@ jacoco {
 
 group = 'uk.gov.hmcts.reform.divorce'
 version = '3.0.0'
-sourceCompatibility = 11
-targetCompatibility = 11
 
-checkstyle {
-    maxWarnings = 0
-    toolVersion = '8.35'
-    // need to set configDir to rootDir otherwise submodule will use submodule/config/checkstyle
-    configDir = new File(rootDir, './')
-    checkstyle.configFile = new File(rootDir, "checkstyle.xml")
+java {
+    toolchain {
+        languageVersion = JavaLanguageVersion.of(17)
+    }
+}
+
+checkstyle.toolVersion = '8.32'
+checkstyle.configFile = new File(rootDir, "checkstyle.xml")
+checkstyle.maxWarnings(0)
+
+// make build fail on Checkstyle issues (https://github.com/gradle/gradle/issues/881)
+tasks.withType(Checkstyle).each { checkstyleTask ->
+    checkstyleTask.doLast {
+        reports.all { report ->
+            def outputFile = report.destination
+            if (outputFile.exists() && outputFile.text.contains("<error ")) {
+                throw new GradleException("There were checkstyle warnings! For more info check $outputFile")
+            }
+        }
+    }
+}
+
+tasks.withType(Test).configureEach {
+    // Using JUnitPlatform for running tests
+    useJUnitPlatform()
 }
 
 bootJar {
