@@ -1,22 +1,21 @@
 package uk.gov.hmcts.reform.divorce.feepayment.test;
 
 import io.restassured.RestAssured;
-import net.serenitybdd.junit.runners.SerenityParameterizedRunner;
 import net.serenitybdd.junit.spring.integration.SpringIntegrationMethodRule;
 import net.thucydides.junit.annotations.TestData;
 import org.junit.Before;
 import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
 import org.springframework.boot.autoconfigure.http.HttpMessageConvertersAutoConfiguration;
-import org.springframework.cloud.netflix.ribbon.RibbonAutoConfiguration;
 import org.springframework.cloud.openfeign.FeignAutoConfiguration;
-import org.springframework.cloud.openfeign.ribbon.FeignRibbonClientAutoConfiguration;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import java.util.Arrays;
 import java.util.Collection;
@@ -26,10 +25,10 @@ import static net.serenitybdd.rest.SerenityRest.when;
 import static org.hamcrest.core.Is.isA;
 
 @Lazy
-@RunWith(SerenityParameterizedRunner.class)
+@ExtendWith(SpringExtension.class)
 @ComponentScan(basePackages = {"uk.gov.hmcts.reform.divorce.feepayment.test", "uk.gov.hmcts.auth.provider.service"})
-@ImportAutoConfiguration({RibbonAutoConfiguration.class, HttpMessageConvertersAutoConfiguration.class,
-        FeignRibbonClientAutoConfiguration.class, FeignAutoConfiguration.class})
+@ImportAutoConfiguration({HttpMessageConvertersAutoConfiguration.class,
+        FeignAutoConfiguration.class})
 @ContextConfiguration(classes = {ServiceContextConfiguration.class})
 public class FeePaymentIntegrationTest {
 
@@ -46,9 +45,6 @@ public class FeePaymentIntegrationTest {
         baseURI = feesPaymentsServiceUrl;
     }
 
-    public FeePaymentIntegrationTest(String feeEndpoint) {
-        this.feeEndpoint = feeEndpoint;
-    }
 
     @TestData
     public static Collection<Object[]> data() {
@@ -63,8 +59,9 @@ public class FeePaymentIntegrationTest {
         });
     }
 
-    @Test
-    public void feeTest() {
+    @ParameterizedTest
+    @MethodSource("data")
+    public void feeTest(String feeEndpoint) {
         RestAssured.useRelaxedHTTPSValidation();
         when()
                 .get(feeEndpoint)
