@@ -2,11 +2,11 @@ package uk.gov.hmcts.reform.divorce.feepayment;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import uk.gov.hmcts.reform.authorisation.healthcheck.ServiceAuthHealthIndicator;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 
 @SpringBootApplication(scanBasePackages = {"uk.gov.hmcts.reform.divorce",
-        "uk.gov.hmcts.reform.logging.appinsights" } ,
-        exclude = {ServiceAuthHealthIndicator.class})
+        "uk.gov.hmcts.reform.logging.appinsights" })
+@EnableConfigurationProperties
 public class FeesPaymentServiceApplication {
 
     public static void main(String[] args) {
