@@ -1,6 +1,6 @@
 package uk.gov.hmcts.reform.divorce.feepayment.controller;
 
-import io.swagger.annotations.ApiOperation;
+import io.swagger.v3.oas.annotations.Operation;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -20,50 +20,50 @@ public class FeePaymentServiceController {
     @Autowired
     private FeePaymentService feePaymentService;
 
-    @ApiOperation(value = "Get Issue fee", tags = {"Fee Lookup"})
+    @Operation(description = "Get Issue fee", tags = {"Fee Lookup"})
     @GetMapping(value = "/version/1/petition-issue-fee", produces = "application/json")
     public ResponseEntity<Fee> lookupFeesForPetitionIssue() {
         log.info("Getting fee for issue");
         return ResponseEntity.ok(feePaymentService.getIssueFee());
     }
 
-    @ApiOperation(value = "Get amend fee", tags = {"Fee Lookup"})
+    @Operation(description = "Get amend fee", tags = {"Fee Lookup"})
     @GetMapping("/version/1/amend-fee")
     public ResponseEntity<Fee> getAmendFee()  {
         return ResponseEntity.ok(feePaymentService.getAmendPetitionFee());
     }
 
-    @ApiOperation(value = "Get Defended petition fee", tags = {"Fee Lookup"})
+    @Operation(description = "Get Defended petition fee", tags = {"Fee Lookup"})
     @GetMapping("/version/1/defended-petition-fee")
     public ResponseEntity<Fee>  getDefendedPetitionFee()  {
         return ResponseEntity.ok(feePaymentService.getDefendPetitionFee());
     }
 
-    @ApiOperation(value = "Get General Application fee", tags = {"Fee Lookup"})
+    @Operation(description = "Get General Application fee", tags = {"Fee Lookup"})
     @GetMapping("/version/1/general-application-fee")
     public ResponseEntity<Fee>  getGeneralApplicationFee()  {
         return ResponseEntity.ok(feePaymentService.getGeneralApplicationFee());
     }
 
-    @ApiOperation(value = "Get Enforcement fee", tags = {"Fee Lookup"})
+    @Operation(description = "Get Enforcement fee", tags = {"Fee Lookup"})
     @GetMapping("/version/1/enforcement-fee")
     public ResponseEntity<Fee>  getEnforcementFee()  {
         return ResponseEntity.ok(feePaymentService.getEnforcementFee());
     }
 
-    @ApiOperation(value = "Get Application Financial Order fee", tags = {"Fee Lookup"})
+    @Operation(description = "Get Application Financial Order fee", tags = {"Fee Lookup"})
     @GetMapping("/version/1/application-financial-order-fee")
     public ResponseEntity<Fee>  getFinancialOrderFee()  {
         return ResponseEntity.ok(feePaymentService.getApplicationFinancialOrderFee());
     }
 
-    @ApiOperation(value = "Get Enforcement fee", tags = {"Fee Lookup"})
+    @Operation(description = "Get Enforcement fee", tags = {"Fee Lookup"})
     @GetMapping("/version/1/application-without-notice-fee")
     public ResponseEntity<Fee>  getApplicationWithoutNoticeFee()  {
         return ResponseEntity.ok(feePaymentService.getApplicationWithoutNoticeFee());
     }
 
-    @ApiOperation(value = "Get all fees", tags = {"Fee Lookup"})
+    @Operation(description = "Get all fees", tags = {"Fee Lookup"})
     @GetMapping("/version/1/get-all-fees")
     public ResponseEntity<List<Fee>> getAllFees() {
         return ResponseEntity.ok(feePaymentService.getAllFees());
