@@ -1,49 +1,21 @@
 package uk.gov.hmcts.reform.divorce.feepayment.management.config;
 
-import org.springframework.beans.factory.annotation.Value;
+import io.swagger.v3.oas.models.OpenAPI;
+import io.swagger.v3.oas.models.info.Info;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-import uk.gov.hmcts.reform.divorce.feepayment.FeesPaymentServiceApplication;
 
 @Configuration
-@EnableSwagger2
+@ConditionalOnProperty(value = "documentation.swagger.enabled", havingValue = "true")
 public class SwaggerConfiguration implements WebMvcConfigurer {
 
-    @Value("${documentation.swagger.enabled}")
-    private boolean swaggerEnabled;
-
     @Bean
-    public Docket api() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .select()
-                .apis(RequestHandlerSelectors.basePackage(FeesPaymentServiceApplication.class.getPackage().getName()))
-                .build()
-                .useDefaultResponseMessages(true)
-                .apiInfo(apiInfo());
-    }
-
-    private ApiInfo apiInfo() {
-        return new ApiInfoBuilder()
-                .title("Fees and Payments Service API")
-                .description("Service to interact and consolidate fees and payments api")
-                .build();
-    }
-
-    @Override
-    public void addResourceHandlers(ResourceHandlerRegistry registry) {
-        if (swaggerEnabled) {
-            registry.addResourceHandler("/swagger-ui.html**")
-                    .addResourceLocations("classpath:/META-INF/resources/swagger-ui.html");
-            registry.addResourceHandler("/webjars/**")
-                    .addResourceLocations("classpath:/META-INF/resources/webjars/");
-        }
+    public OpenAPI springOpenAPI() {
+        return new OpenAPI()
+                .info(new Info().title("Fees and Payments Service API")
+                        .description("Service to interact and consolidate fees and payments api"));
     }
 }
+
